+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.2.0] - 2025-12-27
+
+- Fix: Tooltip text was hard to read on light themes
+- Added support for customizing the status bar displayed information (`ag-usage.statusBarDisplay`)
+
 ## [1.1.0] - 2025-12-26
 
 - Added support for customizing the refresh interval (`ag-usage.refreshInterval`)